requirements.txt
openai
chromadb
mysql-connector-python
slack_sdk
python-dotenv
langchain